Fixed TREE and SHOP constants according to Johan simulations

+++ app/src/main/java/se/kth/projectarbor/project_arbor/StatsTab.java
@@ -84,7 +84,7 @@ public class StatsTab extends Fragment {
 
             // Msgs from Pedometer: steps and distance
             if (intent.getAction().equals(Pedometer.DISTANCE_BROADCAST)) {
-               dist.setText( ""+ extras.getInt("DISTANCE"));
+               dist.setText( ""+ extras.getDouble("DISTANCE"));
             }
             // TODO: Check that receiver message is correct new version of Pedometer is ready
             if (intent.getAction().equals(Pedometer.DISTANCE_BROADCAST)) {
